requirements.txt
asgiref==3.2.10
configobj==5.0.6
Django==3.1
djangorestframework==3.11.1
pycurl==7.43.0.6
pytz==2020.1
six==1.15.0
slick-bitcoinrpc==0.1.4
sqlparse==0.3.1
ujson==3.1.0
